+++ CHANGELOG.md
@@ -11,7 +11,7 @@
 * Removes edit and delete functionality.
 * Adds `SharedPreferencesExplorerScreen`.
 * Supports `SharedPreferencesAsync` and `SharedPreferencesWithCache`.
-* Makes the anchor button draggable.
+* Makes the floating button draggable.
 * Updates minimum supported SDK version to Flutter 3.22/Dart 3.4.
 
 ## 0.0.2